Intermediate p2p server and client
